requirements.txt
Flask
python-dotenv
langchain
langchain-openai
streamlit
chromadb